fixed linting errors, stuck on client side environment variable errors
