+++ .dockerignore
@@ -1,2 +1,5 @@
 .env*
 !.env*.example
+
+node_modules
+.next
