+++ docker-compose.yml
@@ -1,5 +1,8 @@
 services:
   frontend:
+    env_file:
+      - path: ./apps/frontend/.env.local
+        required: false
     build:
       context: ./apps/frontend
       dockerfile: ./Dockerfile.frontend
